add table lock

+++ src/src/main/java/org/unimelb/cis/swen90007sda8/Mappers/BookingMapper.java
@@ -1,6 +1,7 @@
 package org.unimelb.cis.swen90007sda8.Mappers;
 
 import org.unimelb.cis.swen90007sda8.DBConnector.postgresqlConnector;
+import org.unimelb.cis.swen90007sda8.LockManager.lockManager;
 import org.unimelb.cis.swen90007sda8.Models.bookingModel;
 
 import java.sql.ResultSet;
@@ -25,22 +26,31 @@ public class BookingMapper {
                 e.printStackTrace();
                 return;
             }
-            if(oldBooking==0){
-                stmt = "UPDATE timeslots SET numberofshots = numberofshots -1 WHERE id ="+ booking.getTimeSlot().getId()+" AND numberofshots>1;";
-            }else{stmt = "UPDATE timeslots SET numberofshots = numberofshots + 1 WHERE id ="+ oldBooking+";";
-                conn.connect(stmt);
-                stmt = "DELETE FROM bookings WHERE timeslotid ="+oldBooking+" AND email='"+booking.getUser().getEmail()+"';";
-                conn.connect(stmt);
-                stmt = "UPDATE timeslots SET numberofshots = numberofshots -1 WHERE id ="+ booking.getTimeSlot().getId()+" AND numberofshots>1;";
+            Integer targetNumberOfShots = TimeSlotMapper.find(booking.getTimeSlot().getId()).getNumberofshots();
+            if(targetNumberOfShots>0){
+                if(oldBooking==0){
+                    lockManager.lock("timeslots");
+                    stmt = "UPDATE timeslots SET numberofshots = numberofshots -1 WHERE id ="+ booking.getTimeSlot().getId()+" AND numberofshots>0;";
+                    conn.connect(stmt);
+                    lockManager.unlock();
+                    lockManager.lock("bookings");
+                    stmt = "INSERT INTO bookings(email, timeslotid, vaccinename) VALUES (" +"'"+booking.getUser().getEmail()+"'"+','+booking.getTimeSlot().getId()+','+"'"+booking.getTimeSlot().getVaccine().getName()+"'"+");";
+                    conn.connect(stmt);
+                    lockManager.unlock();
+                }else{
+                    lockManager.lock("timeslots");
+                    stmt = "UPDATE timeslots SET numberofshots = numberofshots + 1 WHERE id ="+ oldBooking+";";
+                    conn.connect(stmt);
+                    stmt = "DELETE FROM bookings WHERE timeslotid ="+oldBooking+" AND email='"+booking.getUser().getEmail()+"';";
+                    conn.connect(stmt);
+                    stmt = "UPDATE timeslots SET numberofshots = numberofshots -1 WHERE id ="+ booking.getTimeSlot().getId()+" AND numberofshots>0;";
+                    conn.connect(stmt);
+                    lockManager.unlock();
+                    lockManager.lock("bookings");
+                    stmt = "INSERT INTO bookings(email, timeslotid, vaccinename) VALUES (" +"'"+booking.getUser().getEmail()+"'"+','+booking.getTimeSlot().getId()+','+"'"+booking.getTimeSlot().getVaccine().getName()+"'"+");";
+                    conn.connect(stmt);
+                    lockManager.unlock();
+                }
             }
-        rs = conn.connect(stmt);
-        try{
-            if(rs.next()){
-                stmt = "INSERT INTO bookings(email, timeslotid, vaccinename) VALUES (" +"'"+booking.getUser().getEmail()+"'"+','+booking.getTimeSlot().getId()+','+"'"+booking.getTimeSlot().getVaccine().getName()+"'"+");";
-                conn.connect(stmt);
-            }
-        }catch (SQLException e){
-            e.printStackTrace();
-        }
     }
 }
